+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 dev:
   - add commit hash to startup
   - warn if blobs are no longer available in the beacon node for a block, but continue
+  - do not error when finality response is unavailable
 
 0.9.0:
   - support electra